project/pulsar-producer/requirements.txt
pulsar-client
maven
PyGithub
requests



